requirements.txt
einops==0.8.0
ftfy==6.3.1
iopath==0.1.10
learn2learn==0.2.0
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
Pillow==11.1.0
python-dotenv==1.0.1
pytorchvideo==0.1.5
regex==2024.5.10
rich==13.9.4
setuptools==75.1.0
soundfile==0.12.1
timm==1.0.13
torch==2.4.1
torchaudio==2.4.1
torchvision==0.20.0
tqdm==4.66.4
transformers==4.37.2
